+++ app/scripts/controllers/bridge-status/types.ts
@@ -9,22 +9,12 @@ import {
   NetworkControllerGetStateAction,
 } from '@metamask/network-controller';
 import { AccountsControllerGetSelectedAccountAction } from '@metamask/accounts-controller';
-import { BridgeHistoryItem } from '../../../../shared/types/bridge-status';
+import {
+  BridgeStatusAction,
+  BridgeStatusControllerState,
+} from '../../../../shared/types/bridge-status';
 import { BRIDGE_STATUS_CONTROLLER_NAME } from './constants';
 import BridgeStatusController from './bridge-status-controller';
-
-// All fields need to be types not interfaces, same with their children fields
-// o/w you get a type error
-type SourceChainTxHash = string;
-export type BridgeStatusControllerState = {
-  txHistory: Record<SourceChainTxHash, BridgeHistoryItem>;
-};
-
-export enum BridgeStatusAction {
-  START_POLLING_FOR_BRIDGE_TX_STATUS = 'startPollingForBridgeTxStatus',
-  WIPE_BRIDGE_STATUS = 'wipeBridgeStatus',
-  GET_STATE = 'getState',
-}
 
 type BridgeStatusControllerAction<
   FunctionName extends keyof BridgeStatusController,
